requirements.txt
alembic==1.11.0
amqp==5.1.1
anyio==3.6.2
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ez-setup==0.9
fastapi==0.95.1
h11==0.14.0
idna==3.4
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
